requirements.txt
# SELMA Dependencies
# Core ML
torch>=2.3.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
bitsandbytes>=0.44.0
trl>=0.12.0
datasets>=3.0.0

# Data processing
lxml>=5.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.2.0
tqdm>=4.66.0
pdfplumber>=0.11.0

# Evaluation
scikit-learn>=1.5.0
rouge-score>=0.1.2

# Utilities
pyyaml>=6.0.0
jsonlines>=4.0.0
huggingface-hub>=0.25.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# Synthetic data generation
anthropic>=0.40.0
